+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `@qiskit/qiskit-qasm`: Make QasmError a class
 - `@qiskit/qiskit-qasm`: Use QasmError for jison error handling
+- `@qiskit/qiskit-qasm`: Make parser a member of class Parser
 
 ## [0.9.0] - 2019-05-13